phase-2-todo-app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
openai-agents>=0.0.7
mcp>=1.0.0
python-dotenv>=1.0.0
